Add PuTTY-related entries to README.md
Added entry to the Table-of-Contents as well as instructions for
applying the .reg files found in the repository. Instructions include
directions for importing .reg files as new PuTTY sessions as well as
overwriting the colors for an existing session. Also included
additional recommended PuTTY settings to make the appearance look more
like terminals found on Linux and OS X.