Add dynamic-colors scripts
Some xterm inspired terminals support changing their palette
via OSC 4 and OSC 10,11,12.  This capability is referred to
as "dynamic colors" in the xterm configuration.

This commit generates shell scripts that emit these OSC sequences
and result in being able to change the theme of the terminal
immediately as the script is run.

I'm using this as a convenient way to change the color scheme
dynamically in https://github.com/wez/wezterm

Note that https://github.com/mbadolato/iTerm2-Color-Schemes/pull/156
is a PR that was aiming to do a similar thing, but hasn't had any
review feedback.  This commit takes a different approach and is
integrated with the update_all.py script.