Add wezterm and dynamic-colors configuration (#189)
* fix nondeterministic ordering for xrdb2kitty.py

Running `./update_all.py` resulted in some changed files when
I ran it locally.  Upon inspection, the use of OrderedDict was
defeated by the random order of the hash used to initialize some
of the additional keys.

* Add script to generate wezterm.toml config snippets

https://github.com/wez/wezterm doesn't currently have a convenient way
to toggle themes, so the user will need to copy and paste from the
generated toml file(s) into their wezterm.toml.

* fix xrdb2Termite.py inconsistency

When I ran update_all.py on master, some of the termite configs were
changed.

Some configs were modified in https://github.com/mbadolato/iTerm2-Color-Schemes/pull/188
citing a problem with parsing the colors in quotes.  That PR didn't
update the script that generates those configs, and only changed 2
configuration files.

This commit removes the quoting from all of the configs and
regenerates all of them.

* Add dynamic-colors scripts

Some xterm inspired terminals support changing their palette
via OSC 4 and OSC 10,11,12.  This capability is referred to
as "dynamic colors" in the xterm configuration.

This commit generates shell scripts that emit these OSC sequences
and result in being able to change the theme of the terminal
immediately as the script is run.

I'm using this as a convenient way to change the color scheme
dynamically in https://github.com/wez/wezterm

Note that https://github.com/mbadolato/iTerm2-Color-Schemes/pull/156
is a PR that was aiming to do a similar thing, but hasn't had any
review feedback.  This commit takes a different approach and is
integrated with the update_all.py script.

* improve xrdb file parsing, add missing settings to wezterm configs

* dynamic-colors: set more colors

Use the new xrdb parser class to simplify the logic, and then
set more color settings.