Fix link for Ayu Mirage color profile
The previous URL's last component was
`Ayu+Mirage.itermcolors`. After manually browsing
to the raw color profile ourselves we found
that last component was `Ayu%20Mirage.itermcolors`
instead.

We have just updated the `href` attribute
of the corresponding `<a>` tag so that it points to
the URL we're actually seeing when navigating to
the raw color scheme.